Reduce visibility of some functions
